refactor. miletoneStore 분리
