+++ frontend/src/page/labelPage/LabelPage.tsx
@@ -1,7 +1,7 @@
 import { useState, Suspense } from 'react';
 import styled from 'styled-components';
 import { useSetRecoilState } from 'recoil';
-import { labelMilestoneClickedState } from 'store/labelMilestoneStore';
+import { labelMilestoneClickedState } from 'store/labelMilestoneTabStore';
 import LabelPageHeader from 'page/labelPage/labelPageHeader/LabelPageHeader';
 import LabelTable from 'page/labelPage/labelTable/LabelTable';
 import LabelEditForm from 'page/labelPage/labelEditForm/LabelEditForm';
