+++ frontend/src/store/labelMilestoneTabStore.ts
@@ -1,0 +1,22 @@
+import { atom, selector } from 'recoil';
+//레이블,모달 클릭 감지 및 리셑_____________________________________
+interface labelMilestoneStateType {
+  label: boolean;
+  milestone: boolean;
+}
+
+export const labelMilestoneClickedState = atom<labelMilestoneStateType>({
+  key: 'labelMilestoneState',
+  default: { label: false, milestone: false },
+});
+
+export const resetTabClickedState = selector({
+  key: 'resetTabClickedState',
+  get: ({ get }) => {
+    const tabClickedState = get(labelMilestoneClickedState);
+    return tabClickedState;
+  },
+  set: ({ reset }) => {
+    reset(labelMilestoneClickedState);
+  },
+});
